adding css grid within the Minigame component, testing on Scythe, temporarily not displaying paragraph tags within it
