+++ src/components/App/App.js
@@ -1,6 +1,6 @@
 import React, { useState, useEffect } from "react";
 import "./App.scss";
-import Minigame from "../Minigame";
+import Minigame from "../Minigame/Minigame";
 // import Round from "../Round/Round";
 import Rules from "../Rules";
 import defaultRounds from "../../rounds";
